+++ source/main/zettlr-menu.js
@@ -89,6 +89,13 @@ class ZettlrMenu
                             if(focusedWindow) focusedWindow.webContents.send('message', { 'command': 'import-lang-file' });
                         }
                     },
+                    {
+                        label: trans('menu.import_dict_file'),
+                        click (item, focusedWindow) {
+                            console.log(require('path').join(require('electron').app.getPath('userData'), '/dict'));
+                            require('electron').shell.openItem(require('path').join(require('electron').app.getPath('userData'), '/dict'));
+                        }
+                    },
                     { type: 'separator' },
                     {
                         label: trans('menu.rename_file'),